chapter8/web-application/requirements.txt
gunicorn
tensorflow
flask
matplotlib
Pillow
